+++ account/src/main/java/com/iag/bbscloud/account/model/repository/AccountStatisticRepository.java
@@ -9,5 +9,5 @@ import java.math.BigInteger;
  * @Author xueshan.wei@mljr.com
  * @Date 2017/9/27 上午9:44
  */
-public interface AccountStatisticRepository extends CrudRepository<AccountStatistic, BigInteger> {
+public interface AccountStatisticRepository extends CrudRepository<AccountStatistic, Long> {
 }
